+++ src/constellation/api_devices_block.go
@@ -92,7 +92,7 @@ func DeviceBlock(w http.ResponseWriter, req *http.Request) {
 			}
 			
 			go func() {
-				SendNewDBSyncMessage()
+				go SendNewDBSyncMessage()
 				time.Sleep(2 * time.Second)
 				RestartNebula()
 			}()