+++ src/constellation/api_devices_edit.go
@@ -89,7 +89,7 @@ func DeviceEdit_API(w http.ResponseWriter, req *http.Request) {
 	})
 
 	go func() {
-		SendNewDBSyncMessage()
+		go SendNewDBSyncMessage()
 		time.Sleep(2 * time.Second)
 		RestartNebula()
 	}()