+++ src/utils/types.go
@@ -262,9 +262,9 @@ type ProxyRouteConfig struct {
 	ExtraHeaders               map[string]string           `yaml:"extra_headers,omitempty"`
 	SkipURLClean               bool                        `yaml:"skip_url_clean"`
 	UseH2C                     bool                        `yaml:"use_h2c"`
-	LBMode                     string                      `yaml:"lb_mode" json:"lb_mode,omitempty"`
-	LBStickyMode               bool                        `yaml:"lb_sticky_mode" json:"lb_sticky_mode,omitempty"`
-	Const_IsTunneled           bool                        `yaml:"-", json:"-"`
+	LBMode                     string                      `yaml:"lb_mode" json:"LBMode,omitempty"`
+	LBStickyMode               bool                        `yaml:"lb_sticky_mode" json:"LBStickyMode,omitempty"`
+	Const_IsTunneled           bool                        `yaml:"-" json:"-"`
 }
 
 type EmailConfig struct {